Move utility files to a utility folder
