+++ SolarSense/src/utility/priceZoneMapper.ts
