+++ SolarSense/src/utility/solarFetch.tsx
